refactor(api): reorganize modules
