+++ packages/backend/src/photos/api.ts
@@ -1,48 +1,12 @@
 import {TypeBoxTypeProvider} from '@fastify/type-provider-typebox';
-import {Static, Type} from '@sinclair/typebox';
 import Fastify from 'fastify';
-import {PhotoCommandOutput, getPhotos} from '../db/photosTable';
+import {getPhotos} from '../db/photosTable';
+import {parseDbOutput} from './parseDbOutput';
+import {PhotosReply, PhotosReplyType} from './types';
 
 const server = Fastify({
   logger: process.env.NODE_ENV !== 'test',
 }).withTypeProvider<TypeBoxTypeProvider>();
-
-const PhotosReply = Type.Array(
-  Type.Object({
-    id: Type.String(),
-    category: Type.String(),
-    orderCount: Type.Number(),
-    extra: Type.Optional(
-      Type.Partial(
-        Type.Object({
-          texture: Type.String(),
-          border: Type.Number(),
-          rotate: Type.Number(),
-        })
-      )
-    ),
-  })
-);
-
-export type PhotosReplyType = Static<typeof PhotosReply>;
-
-// parse db output into appropriate API response
-const parseDbOutputPhotos = (
-  photos: PhotoCommandOutput['Items'] = []
-): PhotosReplyType => {
-  return photos.map(photo => {
-    return {
-      id: photo.id.S,
-      category: photo.category.S,
-      orderCount: photo.orderCount.N,
-      extra: {
-        border: photo.extra?.M.border?.N,
-        texture: photo.extra?.M.texture?.S,
-        rotate: photo.extra?.M.rotate?.N,
-      },
-    };
-  });
-};
 
 server.get<{Reply: PhotosReplyType}>(
   '/photos',
@@ -55,7 +19,7 @@ server.get<{Reply: PhotosReplyType}>(
   },
   async () => {
     const {Items = []} = await getPhotos();
-    return parseDbOutputPhotos(Items);
+    return parseDbOutput(Items);
   }
 );
